change to int

+++ Likelihood/src/WcsMap.cxx
@@ -147,8 +147,8 @@ double WcsMap::operator()(const astro::SkyDir & dir) const {
    double x(pixel.first);
    double y(pixel.second);
 
-   size_t ix = static_cast<size_t>(x);
-   size_t iy = static_cast<size_t>(y);
+   int ix = static_cast<int>(x);
+   int iy = static_cast<int>(y);
 
 // NB: wcslib starts indexing pixels with 1, not 0.
    if (ix < 1) {
